requirements.txt
jira==3.5.1
python-gitlab==3.15.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
responses==0.24.1
openai==1.3.5
pyyaml==6.0.1 